requirements.txt
torch~=2.4.0+cu118
matplotlib~=3.9.1
numpy~=1.23.1
scipy~=1.13.1
tqdm~=4.66.5
einops~=0.8.0
open3d~=0.18.0
pillow~=10.2.0